Remove CustomText.selectable
