+++ example/README.md
@@ -9,5 +9,4 @@ A demo app showing all of the following examples.
 - Example 5 - [SelectiveDefinition](https://github.com/kaboc/flutter_custom_text/blob/main/example/lib/examples/example5.dart)
 - Example 6 - [SpanDefinition](https://github.com/kaboc/flutter_custom_text/blob/main/example/lib/examples/example6.dart)
 - Example 7 - [Changing mouse cursor and text style on hover](https://github.com/kaboc/flutter_custom_text/blob/main/example/lib/examples/example7.dart)
-- Example 8 - [CustomText.selectable](https://github.com/kaboc/flutter_custom_text/blob/main/example/lib/examples/example8.dart)
-- Example 9 - [CustomTextEditingController](https://github.com/kaboc/flutter_custom_text/blob/main/example/lib/examples/example9.dart)
+- Example 8 - [CustomTextEditingController](https://github.com/kaboc/flutter_custom_text/blob/main/example/lib/examples/example8.dart)
