+++ lib/src/text.dart
@@ -1,7 +1,5 @@
 import 'dart:async';
-import 'dart:ui' as ui show BoxHeightStyle, BoxWidthStyle;
 
-import 'package:flutter/gestures.dart' show DragStartBehavior;
 import 'package:flutter/material.dart';
 
 import 'package:text_parser/text_parser.dart';
@@ -48,74 +46,7 @@ class CustomText extends StatefulWidget {
     this.semanticsLabel,
     this.textWidthBasis,
     this.textHeightBehavior,
-  })  : _isSelectable = false,
-        focusNode = null,
-        showCursor = false,
-        autofocus = false,
-        toolbarOptions = null,
-        minLines = null,
-        cursorWidth = 0.0,
-        cursorHeight = null,
-        cursorRadius = null,
-        cursorColor = null,
-        selectionHeightStyle = null,
-        selectionWidthStyle = null,
-        dragStartBehavior = null,
-        enableInteractiveSelection = false,
-        selectionControls = null,
-        scrollPhysics = null,
-        onSelectionChanged = null;
-
-  /// Creates a selectable text widget that decorates strings in it
-  /// and enables tap, long-press and/or hover gestures based on
-  /// flexible definitions.
-  @Deprecated(
-    'Use SelectionArea on Flutter 3.3 and above, or '
-    '`TextField.readOnly` with `CustomTextEditingController`',
-  )
-  const CustomText.selectable(
-    this.text, {
-    super.key,
-    required this.definitions,
-    this.parserOptions = const ParserOptions(),
-    this.style,
-    this.matchStyle,
-    this.tapStyle,
-    this.hoverStyle,
-    this.onTap,
-    this.onLongPress,
-    this.longPressDuration,
-    this.preventBlocking = false,
-    this.focusNode,
-    this.strutStyle,
-    this.textAlign,
-    this.textDirection,
-    this.textScaleFactor,
-    this.showCursor = false,
-    this.autofocus = false,
-    this.toolbarOptions,
-    this.minLines,
-    this.maxLines,
-    this.cursorWidth = 2.0,
-    this.cursorHeight,
-    this.cursorRadius,
-    this.cursorColor,
-    this.selectionHeightStyle,
-    this.selectionWidthStyle,
-    this.dragStartBehavior,
-    this.enableInteractiveSelection = true,
-    this.selectionControls,
-    this.scrollPhysics,
-    this.semanticsLabel,
-    this.textHeightBehavior,
-    this.textWidthBasis,
-    this.onSelectionChanged,
-  })  : _isSelectable = true,
-        locale = null,
-        softWrap = null,
-        overflow = null;
-
-  final bool _isSelectable;
+  });
 
   /// The text to parse and display.
   final String text;
@@ -213,25 +144,6 @@ class CustomText extends StatefulWidget {
   final String? semanticsLabel;
   final TextWidthBasis? textWidthBasis;
   final TextHeightBehavior? textHeightBehavior;
-
-  // For SelectableText
-  final FocusNode? focusNode;
-  final bool showCursor;
-  final bool autofocus;
-  // ignore: deprecated_member_use
-  final ToolbarOptions? toolbarOptions;
-  final int? minLines;
-  final double cursorWidth;
-  final double? cursorHeight;
-  final Radius? cursorRadius;
-  final Color? cursorColor;
-  final ui.BoxHeightStyle? selectionHeightStyle;
-  final ui.BoxWidthStyle? selectionWidthStyle;
-  final DragStartBehavior? dragStartBehavior;
-  final bool enableInteractiveSelection;
-  final TextSelectionControls? selectionControls;
-  final ScrollPhysics? scrollPhysics;
-  final SelectionChangedCallback? onSelectionChanged;
 
   @override
   State<CustomText> createState() => _CustomTextState();
@@ -350,58 +262,20 @@ class _CustomTextState extends State<CustomText> {
   Widget build(BuildContext context) {
     return ValueListenableBuilder<TextSpan>(
       valueListenable: _textSpanNotifier,
-      builder: (context, span, _) =>
-          widget._isSelectable ? _richSelectableText(span) : _richText(span),
-    );
-  }
-
-  Text _richText(TextSpan span) {
-    return Text.rich(
-      span,
-      strutStyle: widget.strutStyle,
-      textAlign: widget.textAlign,
-      textDirection: widget.textDirection,
-      locale: widget.locale,
-      softWrap: widget.softWrap,
-      overflow: widget.overflow,
-      textScaleFactor: widget.textScaleFactor,
-      maxLines: widget.maxLines,
-      semanticsLabel: widget.semanticsLabel,
-      textWidthBasis: widget.textWidthBasis,
-      textHeightBehavior: widget.textHeightBehavior,
-    );
-  }
-
-  SelectableText _richSelectableText(TextSpan span) {
-    return SelectableText.rich(
-      span,
-      focusNode: widget.focusNode,
-      style: widget.style,
-      strutStyle: widget.strutStyle,
-      textAlign: widget.textAlign,
-      textDirection: widget.textDirection,
-      textScaleFactor: widget.textScaleFactor,
-      showCursor: widget.showCursor,
-      autofocus: widget.autofocus,
-      // ignore: deprecated_member_use
-      toolbarOptions: widget.toolbarOptions,
-      minLines: widget.minLines,
-      maxLines: widget.maxLines,
-      cursorWidth: widget.cursorWidth,
-      cursorHeight: widget.cursorHeight,
-      cursorRadius: widget.cursorRadius,
-      cursorColor: widget.cursorColor,
-      selectionHeightStyle:
-          widget.selectionHeightStyle ?? ui.BoxHeightStyle.tight,
-      selectionWidthStyle: widget.selectionWidthStyle ?? ui.BoxWidthStyle.tight,
-      dragStartBehavior: widget.dragStartBehavior ?? DragStartBehavior.start,
-      enableInteractiveSelection: widget.enableInteractiveSelection,
-      selectionControls: widget.selectionControls,
-      scrollPhysics: widget.scrollPhysics,
-      semanticsLabel: widget.semanticsLabel,
-      textHeightBehavior: widget.textHeightBehavior,
-      textWidthBasis: widget.textWidthBasis,
-      onSelectionChanged: widget.onSelectionChanged,
+      builder: (context, span, _) => Text.rich(
+        span,
+        strutStyle: widget.strutStyle,
+        textAlign: widget.textAlign,
+        textDirection: widget.textDirection,
+        locale: widget.locale,
+        softWrap: widget.softWrap,
+        overflow: widget.overflow,
+        textScaleFactor: widget.textScaleFactor,
+        maxLines: widget.maxLines,
+        semanticsLabel: widget.semanticsLabel,
+        textWidthBasis: widget.textWidthBasis,
+        textHeightBehavior: widget.textHeightBehavior,
+      ),
     );
   }
 }
