+++ test/utils.dart
@@ -24,19 +24,9 @@ Text findText() {
   return finder.evaluate().first.widget as Text;
 }
 
-SelectableText findSelectableText() {
-  final finder = find.byType(SelectableText);
-  return finder.evaluate().first.widget as SelectableText;
-}
-
 RichText findRichText() {
   final finder = find.byType(RichText);
   return finder.evaluate().first.widget as RichText;
-}
-
-EditableText findEditableText() {
-  final finder = find.byType(EditableText);
-  return finder.evaluate().first.widget as EditableText;
 }
 
 List<InlineSpan> getSpans() {
@@ -48,32 +38,9 @@ List<InlineSpan> getSpans() {
   return spans;
 }
 
-List<InlineSpan> getSelectableSpans() {
-  final spans = <InlineSpan>[];
-  findSelectableText().textSpan?.visitChildren((span) {
-    spans.add(span);
-    return true;
-  });
-  return spans;
-}
-
 InlineSpan? findSpan(String text) {
   InlineSpan? span;
   findText().textSpan?.visitChildren(
-    (visitor) {
-      if (visitor is TextSpan && visitor.text == text) {
-        span = visitor;
-        return false;
-      }
-      return true;
-    },
-  );
-  return span;
-}
-
-InlineSpan? findSelectableSpan(String text) {
-  InlineSpan? span;
-  findSelectableText().textSpan?.visitChildren(
     (visitor) {
       if (visitor is TextSpan && visitor.text == text) {
         span = visitor;
